Plots coverage with classification results

This function is used for visualization of the classification method for peptides. It uses the traditional form of coverage, with length present and easily estimated redundancy. It also allows to see if the tendencies for exchange are coherent in regions.

Usage

plot_cov_class(fit_values, fractional = TRUE, interactive = FALSE)

Arguments

fit_values

...

fractional

...

interactive

...

Value

a girafe object.

See also

Examples

kin_dat <- prepare_kin_dat(alpha_dat)
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values <- create_fit_dataset(kin_dat, fit_k_params, control)
plot_cov_class(fit_values)