Prepare coloring commands for ChimeraX

This function prepares the coloring commands for ChimeraX. If you wish only to create coloring commands and downlaod them as a text file, see export_chimera_commands function.

Usage

prepare_chimera_commands(
  hires_params,
  dominant = FALSE,
  state = hires_params[["State"]][1],
  chains = c("A")
)

Arguments

state

biological state

chain

chain for which the coloring commands are created

Examples

fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat = kin_dat, control = control, fit_k_params = fit_k_params, fractional = TRUE)
hires_params <- calculate_hires(fit_values, method = "weighted", fractional = TRUE)

prepare_chimera_commands(hires_params, chains = c("A", "G"))
#>   [1] "color /A:1 #7A5233 color /G:1 #7A5233 "    
#>   [2] "color /A:2 #8C5B18 color /G:2 #8C5B18 "    
#>   [3] "color /A:3 #915F0F color /G:3 #915F0F "    
#>   [4] "color /A:4 #915F0F color /G:4 #915F0F "    
#>   [5] "color /A:5 #877107 color /G:5 #877107 "    
#>   [6] "color /A:6 #877107 color /G:6 #877107 "    
#>   [7] "color /A:7 #877107 color /G:7 #877107 "    
#>   [8] "color /A:8 #000000 color /G:8 #000000 "    
#>   [9] "color /A:9 #877107 color /G:9 #877107 "    
#>  [10] "color /A:10 #877107 color /G:10 #877107 "  
#>  [11] "color /A:11 #8A6B0A color /G:11 #8A6B0A "  
#>  [12] "color /A:12 #1600E9 color /G:12 #1600E9 "  
#>  [13] "color /A:13 #1600E9 color /G:13 #1600E9 "  
#>  [14] "color /A:14 #1600E9 color /G:14 #1600E9 "  
#>  [15] "color /A:15 #1800E7 color /G:15 #1800E7 "  
#>  [16] "color /A:16 #1800E7 color /G:16 #1800E7 "  
#>  [17] "color /A:17 #2411CA color /G:17 #2411CA "  
#>  [18] "color /A:18 #2516C4 color /G:18 #2516C4 "  
#>  [19] "color /A:19 #2916C0 color /G:19 #2916C0 "  
#>  [20] "color /A:20 #2916C0 color /G:20 #2916C0 "  
#>  [21] "color /A:21 #3319B3 color /G:21 #3319B3 "  
#>  [22] "color /A:22 #49169F color /G:22 #49169F "  
#>  [23] "color /A:23 #560F9A color /G:23 #560F9A "  
#>  [24] "color /A:24 #620894 color /G:24 #620894 "  
#>  [25] "color /A:25 #620894 color /G:25 #620894 "  
#>  [26] "color /A:26 #620894 color /G:26 #620894 "  
#>  [27] "color /A:27 #5A1095 color /G:27 #5A1095 "  
#>  [28] "color /A:28 #5A1095 color /G:28 #5A1095 "  
#>  [29] "color /A:29 #000000 color /G:29 #000000 "  
#>  [30] "color /A:30 #5A1095 color /G:30 #5A1095 "  
#>  [31] "color /A:31 #5A1095 color /G:31 #5A1095 "  
#>  [32] "color /A:32 #5A1095 color /G:32 #5A1095 "  
#>  [33] "color /A:33 #5A1095 color /G:33 #5A1095 "  
#>  [34] "color /A:34 #0000FF color /G:34 #0000FF "  
#>  [35] "color /A:35 #0000FF color /G:35 #0000FF "  
#>  [36] "color /A:36 #0000FF color /G:36 #0000FF "  
#>  [37] "color /A:37 #3900C6 color /G:37 #3900C6 "  
#>  [38] "color /A:38 #3900C6 color /G:38 #3900C6 "  
#>  [39] "color /A:39 #95006A color /G:39 #95006A "  
#>  [40] "color /A:40 #851169 color /G:40 #851169 "  
#>  [41] "color /A:41 #851169 color /G:41 #851169 "  
#>  [42] "color /A:42 #851169 color /G:42 #851169 "  
#>  [43] "color /A:43 #000000 color /G:43 #000000 "  
#>  [44] "color /A:44 #000000 color /G:44 #000000 "  
#>  [45] "color /A:45 #000000 color /G:45 #000000 "  
#>  [46] "color /A:46 #851169 color /G:46 #851169 "  
#>  [47] "color /A:47 #851169 color /G:47 #851169 "  
#>  [48] "color /A:48 #772068 color /G:48 #772068 "  
#>  [49] "color /A:49 #772068 color /G:49 #772068 "  
#>  [50] "color /A:50 #772068 color /G:50 #772068 "  
#>  [51] "color /A:51 #1000EF color /G:51 #1000EF "  
#>  [52] "color /A:52 #0800F7 color /G:52 #0800F7 "  
#>  [53] "color /A:53 #0800F7 color /G:53 #0800F7 "  
#>  [54] "color /A:54 #0800F7 color /G:54 #0800F7 "  
#>  [55] "color /A:55 #0800F7 color /G:55 #0800F7 "  
#>  [56] "color /A:56 #0800F7 color /G:56 #0800F7 "  
#>  [57] "color /A:57 #0800F7 color /G:57 #0800F7 "  
#>  [58] "color /A:58 #E20716 color /G:58 #E20716 "  
#>  [59] "color /A:59 #E20716 color /G:59 #E20716 "  
#>  [60] "color /A:60 #E20716 color /G:60 #E20716 "  
#>  [61] "color /A:61 #E20716 color /G:61 #E20716 "  
#>  [62] "color /A:62 #E20716 color /G:62 #E20716 "  
#>  [63] "color /A:63 #E20716 color /G:63 #E20716 "  
#>  [64] "color /A:64 #F20409 color /G:64 #F20409 "  
#>  [65] "color /A:65 #F20409 color /G:65 #F20409 "  
#>  [66] "color /A:66 #F40407 color /G:66 #F40407 "  
#>  [67] "color /A:67 #F40407 color /G:67 #F40407 "  
#>  [68] "color /A:68 #F40407 color /G:68 #F40407 "  
#>  [69] "color /A:69 #000000 color /G:69 #000000 "  
#>  [70] "color /A:70 #F40407 color /G:70 #F40407 "  
#>  [71] "color /A:71 #F40407 color /G:71 #F40407 "  
#>  [72] "color /A:72 #F40407 color /G:72 #F40407 "  
#>  [73] "color /A:73 #F40407 color /G:73 #F40407 "  
#>  [74] "color /A:74 #F90303 color /G:74 #F90303 "  
#>  [75] "color /A:75 #F90303 color /G:75 #F90303 "  
#>  [76] "color /A:76 #F90303 color /G:76 #F90303 "  
#>  [77] "color /A:77 #F90303 color /G:77 #F90303 "  
#>  [78] "color /A:78 #F90303 color /G:78 #F90303 "  
#>  [79] "color /A:79 #FC0200 color /G:79 #FC0200 "  
#>  [80] "color /A:80 #000000 color /G:80 #000000 "  
#>  [81] "color /A:81 #FC0200 color /G:81 #FC0200 "  
#>  [82] "color /A:82 #FC0200 color /G:82 #FC0200 "  
#>  [83] "color /A:83 #FE0001 color /G:83 #FE0001 "  
#>  [84] "color /A:84 #FE0001 color /G:84 #FE0001 "  
#>  [85] "color /A:85 #FF0000 color /G:85 #FF0000 "  
#>  [86] "color /A:86 #FF0000 color /G:86 #FF0000 "  
#>  [87] "color /A:87 #FF0000 color /G:87 #FF0000 "  
#>  [88] "color /A:88 #FF0000 color /G:88 #FF0000 "  
#>  [89] "color /A:89 #FF0000 color /G:89 #FF0000 "  
#>  [90] "color /A:90 #FF0000 color /G:90 #FF0000 "  
#>  [91] "color /A:91 #FF0000 color /G:91 #FF0000 "  
#>  [92] "color /A:92 #FF0000 color /G:92 #FF0000 "  
#>  [93] "color /A:93 #FF0000 color /G:93 #FF0000 "  
#>  [94] "color /A:94 #FF0000 color /G:94 #FF0000 "  
#>  [95] "color /A:95 #FF0000 color /G:95 #FF0000 "  
#>  [96] "color /A:96 #FF0000 color /G:96 #FF0000 "  
#>  [97] "color /A:97 #FF0000 color /G:97 #FF0000 "  
#>  [98] "color /A:98 #FF0000 color /G:98 #FF0000 "  
#>  [99] "color /A:99 #FF0000 color /G:99 #FF0000 "  
#> [100] "color /A:100 #FF0000 color /G:100 #FF0000 "
#> [101] "color /A:101 #FF0000 color /G:101 #FF0000 "
#> [102] "color /A:102 #FF0000 color /G:102 #FF0000 "
#> [103] "color /A:103 #FD0200 color /G:103 #FD0200 "
#> [104] "color /A:104 #FD0200 color /G:104 #FD0200 "
#> [105] "color /A:105 #FD0200 color /G:105 #FD0200 "
#> [106] "color /A:106 #FD0200 color /G:106 #FD0200 "
#> [107] "color /A:107 #FD0200 color /G:107 #FD0200 "
#> [108] "color /A:108 #FF0000 color /G:108 #FF0000 "
#> [109] "color /A:109 #FF0000 color /G:109 #FF0000 "
#> [110] "color /A:110 #FF0000 color /G:110 #FF0000 "
#> [111] "color /A:111 #FF0000 color /G:111 #FF0000 "
#> [112] "color /A:112 #FF0000 color /G:112 #FF0000 "
#> [113] "color /A:113 #FF0000 color /G:113 #FF0000 "
#> [114] "color /A:114 #FF0000 color /G:114 #FF0000 "
#> [115] "color /A:115 #FF0000 color /G:115 #FF0000 "
#> [116] "color /A:116 #FF0000 color /G:116 #FF0000 "
#> [117] "color /A:117 #FF0000 color /G:117 #FF0000 "
#> [118] "color /A:118 #FF0000 color /G:118 #FF0000 "
#> [119] "color /A:119 #FF0000 color /G:119 #FF0000 "
#> [120] "color /A:120 #FF0000 color /G:120 #FF0000 "
#> [121] "color /A:121 #FF0000 color /G:121 #FF0000 "
#> [122] "color /A:122 #FF0000 color /G:122 #FF0000 "
#> [123] "color /A:123 #FF0000 color /G:123 #FF0000 "
#> [124] "color /A:124 #FF0000 color /G:124 #FF0000 "
#> [125] "color /A:125 #FF0000 color /G:125 #FF0000 "
#> [126] "color /A:126 #FF0000 color /G:126 #FF0000 "
#> [127] "color /A:127 #FF0000 color /G:127 #FF0000 "
#> [128] "color /A:128 #FF0000 color /G:128 #FF0000 "
#> [129] "color /A:129 #FF0000 color /G:129 #FF0000 "
#> [130] "color /A:130 #FF0000 color /G:130 #FF0000 "
#> [131] "color /A:131 #FF0000 color /G:131 #FF0000 "
#> [132] "color /A:132 #FF0000 color /G:132 #FF0000 "
#> [133] "color /A:133 #000000 color /G:133 #000000 "
#> [134] "color /A:134 #FF0000 color /G:134 #FF0000 "
#> [135] "color /A:135 #FF0000 color /G:135 #FF0000 "
#> [136] "color /A:136 #684354 color /G:136 #684354 "
#> [137] "color /A:137 #714846 color /G:137 #714846 "
#> [138] "color /A:138 #714846 color /G:138 #714846 "
#> [139] "color /A:139 #714846 color /G:139 #714846 "
#> [140] "color /A:140 #714846 color /G:140 #714846 "
#> [141] "color /A:141 #714846 color /G:141 #714846 "
#> [142] "color /A:142 #714846 color /G:142 #714846 "
#> [143] "color /A:143 #5E257C color /G:143 #5E257C "
#> [144] "color /A:144 #711B73 color /G:144 #711B73 "
#> [145] "color /A:145 #711B73 color /G:145 #711B73 "
#> [146] "color /A:146 #711B73 color /G:146 #711B73 "
#> [147] "color /A:147 #000000 color /G:147 #000000 "
#> [148] "color /A:148 #711B73 color /G:148 #711B73 "
#> [149] "color /A:149 #711B73 color /G:149 #711B73 "
#> [150] "color /A:150 #711B73 color /G:150 #711B73 "
#> [151] "color /A:151 #8D1B57 color /G:151 #8D1B57 "
#> [152] "color /A:152 #8D1B57 color /G:152 #8D1B57 "
#> [153] "color /A:153 #8D1B57 color /G:153 #8D1B57 "
#> [154] "color /A:154 #433A82 color /G:154 #433A82 "
#> [155] "color /A:155 #0D27CB color /G:155 #0D27CB "
#> [156] "color /A:156 #0D27CB color /G:156 #0D27CB "
#> [157] "color /A:157 #0D27CB color /G:157 #0D27CB "
#> [158] "color /A:158 #0D27CB color /G:158 #0D27CB "
#> [159] "color /A:159 #0D27CB color /G:159 #0D27CB "
#> [160] "color /A:160 #773355 color /G:160 #773355 "
#> [161] "color /A:161 #7F3051 color /G:161 #7F3051 "
#> [162] "color /A:162 #7F3051 color /G:162 #7F3051 "
#> [163] "color /A:163 #8E2749 color /G:163 #8E2749 "
#> [164] "color /A:164 #8E2749 color /G:164 #8E2749 "
#> [165] "color /A:165 #8E2749 color /G:165 #8E2749 "
#> [166] "color /A:166 #8D284A color /G:166 #8D284A "
#> [167] "color /A:167 #92313B color /G:167 #92313B "
#> [168] "color /A:168 #92313B color /G:168 #92313B "
#> [169] "color /A:169 #92323B color /G:169 #92323B "
#> [170] "color /A:170 #874137 color /G:170 #874137 "
#> [171] "color /A:171 #874137 color /G:171 #874137 "
#> [172] "color /A:172 #874137 color /G:172 #874137 "
#> [173] "color /A:173 #874137 color /G:173 #874137 "
#> [174] "color /A:174 #874137 color /G:174 #874137 "
#> [175] "color /A:175 #864832 color /G:175 #864832 "
#> [176] "color /A:176 #864832 color /G:176 #864832 "
#> [177] "color /A:177 #724F3D color /G:177 #724F3D "
#> [178] "color /A:178 #000000 color /G:178 #000000 "
#> [179] "color /A:179 #724F3D color /G:179 #724F3D "
#> [180] "color /A:180 #724F3D color /G:180 #724F3D "
#> [181] "color /A:181 #724F3D color /G:181 #724F3D "
#> [182] "color /A:182 #695442 color /G:182 #695442 "
#> [183] "color /A:183 #625746 color /G:183 #625746 "
#> [184] "color /A:184 #625746 color /G:184 #625746 "
#> [185] "color /A:185 #375B6D color /G:185 #375B6D "
#> [186] "color /A:186 #2F6170 color /G:186 #2F6170 "
#> [187] "color /A:187 #2F6170 color /G:187 #2F6170 "
#> [188] "color /A:188 #2F6170 color /G:188 #2F6170 "
#> [189] "color /A:189 #2F6170 color /G:189 #2F6170 "
#> [190] "color /A:190 #445E5D color /G:190 #445E5D "
#> [191] "color /A:191 #746823 color /G:191 #746823 "
#> [192] "color /A:192 #746823 color /G:192 #746823 "
#> [193] "color /A:193 #746823 color /G:193 #746823 "
#> [194] "color /A:194 #7E6918 color /G:194 #7E6918 "
#> [195] "color /A:195 #8A6212 color /G:195 #8A6212 "
#> [196] "color /A:196 #8A6212 color /G:196 #8A6212 "
#> [197] "color /A:197 #92521B color /G:197 #92521B "
#> [198] "color /A:198 #92521B color /G:198 #92521B "
#> [199] "color /A:199 #92521B color /G:199 #92521B "
#> [200] "color /A:200 #92521B color /G:200 #92521B "
#> [201] "color /A:201 #92521B color /G:201 #92521B "
#> [202] "color /A:202 #556545 color /G:202 #556545 "
#> [203] "color /A:203 #0B8470 color /G:203 #0B8470 "
#> [204] "color /A:204 #03807C color /G:204 #03807C "
#> [205] "color /A:205 #028E6F color /G:205 #028E6F "
#> [206] "color /A:206 #028E6F color /G:206 #028E6F "
#> [207] "color /A:207 #028E6F color /G:207 #028E6F "
#> [208] "color /A:208 #028C71 color /G:208 #028C71 "
#> [209] "color /A:209 #0A866F color /G:209 #0A866F "
#> [210] "color /A:210 #395D69 color /G:210 #395D69 "
#> [211] "color /A:211 #4A3D79 color /G:211 #4A3D79 "
#> [212] "color /A:212 #403F7F color /G:212 #403F7F "
#> [213] "color /A:213 #403F7F color /G:213 #403F7F "
#> [214] "color /A:214 #403F7F color /G:214 #403F7F "
#> [215] "color /A:215 #403F7F color /G:215 #403F7F "
#> [216] "color /A:216 #403F7F color /G:216 #403F7F "
#> [217] "color /A:217 #555E4C color /G:217 #555E4C "
#> [218] "color /A:218 #81532B color /G:218 #81532B "
#> [219] "color /A:219 #965C0E color /G:219 #965C0E "
#> [220] "color /A:220 #965C0E color /G:220 #965C0E "
#> [221] "color /A:221 #965C0E color /G:221 #965C0E "
prepare_chimera_commands(hires_params, chain = "G", dominant = TRUE)
#>   [1] "color /G:1 #7A5233 "   "color /G:2 #8C5B18 "   "color /G:3 #915F0F "  
#>   [4] "color /G:4 #915F0F "   "color /G:5 #877107 "   "color /G:6 #877107 "  
#>   [7] "color /G:7 #877107 "   "color /G:8 #000000 "   "color /G:9 #877107 "  
#>  [10] "color /G:10 #877107 "  "color /G:11 #8A6B0A "  "color /G:12 #1600E9 " 
#>  [13] "color /G:13 #1600E9 "  "color /G:14 #1600E9 "  "color /G:15 #1800E7 " 
#>  [16] "color /G:16 #1800E7 "  "color /G:17 #2411CA "  "color /G:18 #2516C4 " 
#>  [19] "color /G:19 #2916C0 "  "color /G:20 #2916C0 "  "color /G:21 #3319B3 " 
#>  [22] "color /G:22 #49169F "  "color /G:23 #560F9A "  "color /G:24 #620894 " 
#>  [25] "color /G:25 #620894 "  "color /G:26 #620894 "  "color /G:27 #5A1095 " 
#>  [28] "color /G:28 #5A1095 "  "color /G:29 #000000 "  "color /G:30 #5A1095 " 
#>  [31] "color /G:31 #5A1095 "  "color /G:32 #5A1095 "  "color /G:33 #5A1095 " 
#>  [34] "color /G:34 #0000FF "  "color /G:35 #0000FF "  "color /G:36 #0000FF " 
#>  [37] "color /G:37 #3900C6 "  "color /G:38 #3900C6 "  "color /G:39 #95006A " 
#>  [40] "color /G:40 #851169 "  "color /G:41 #851169 "  "color /G:42 #851169 " 
#>  [43] "color /G:43 #000000 "  "color /G:44 #000000 "  "color /G:45 #000000 " 
#>  [46] "color /G:46 #851169 "  "color /G:47 #851169 "  "color /G:48 #772068 " 
#>  [49] "color /G:49 #772068 "  "color /G:50 #772068 "  "color /G:51 #1000EF " 
#>  [52] "color /G:52 #0800F7 "  "color /G:53 #0800F7 "  "color /G:54 #0800F7 " 
#>  [55] "color /G:55 #0800F7 "  "color /G:56 #0800F7 "  "color /G:57 #0800F7 " 
#>  [58] "color /G:58 #E20716 "  "color /G:59 #E20716 "  "color /G:60 #E20716 " 
#>  [61] "color /G:61 #E20716 "  "color /G:62 #E20716 "  "color /G:63 #E20716 " 
#>  [64] "color /G:64 #F20409 "  "color /G:65 #F20409 "  "color /G:66 #F40407 " 
#>  [67] "color /G:67 #F40407 "  "color /G:68 #F40407 "  "color /G:69 #000000 " 
#>  [70] "color /G:70 #F40407 "  "color /G:71 #F40407 "  "color /G:72 #F40407 " 
#>  [73] "color /G:73 #F40407 "  "color /G:74 #F90303 "  "color /G:75 #F90303 " 
#>  [76] "color /G:76 #F90303 "  "color /G:77 #F90303 "  "color /G:78 #F90303 " 
#>  [79] "color /G:79 #FC0200 "  "color /G:80 #000000 "  "color /G:81 #FC0200 " 
#>  [82] "color /G:82 #FC0200 "  "color /G:83 #FE0001 "  "color /G:84 #FE0001 " 
#>  [85] "color /G:85 #FF0000 "  "color /G:86 #FF0000 "  "color /G:87 #FF0000 " 
#>  [88] "color /G:88 #FF0000 "  "color /G:89 #FF0000 "  "color /G:90 #FF0000 " 
#>  [91] "color /G:91 #FF0000 "  "color /G:92 #FF0000 "  "color /G:93 #FF0000 " 
#>  [94] "color /G:94 #FF0000 "  "color /G:95 #FF0000 "  "color /G:96 #FF0000 " 
#>  [97] "color /G:97 #FF0000 "  "color /G:98 #FF0000 "  "color /G:99 #FF0000 " 
#> [100] "color /G:100 #FF0000 " "color /G:101 #FF0000 " "color /G:102 #FF0000 "
#> [103] "color /G:103 #FD0200 " "color /G:104 #FD0200 " "color /G:105 #FD0200 "
#> [106] "color /G:106 #FD0200 " "color /G:107 #FD0200 " "color /G:108 #FF0000 "
#> [109] "color /G:109 #FF0000 " "color /G:110 #FF0000 " "color /G:111 #FF0000 "
#> [112] "color /G:112 #FF0000 " "color /G:113 #FF0000 " "color /G:114 #FF0000 "
#> [115] "color /G:115 #FF0000 " "color /G:116 #FF0000 " "color /G:117 #FF0000 "
#> [118] "color /G:118 #FF0000 " "color /G:119 #FF0000 " "color /G:120 #FF0000 "
#> [121] "color /G:121 #FF0000 " "color /G:122 #FF0000 " "color /G:123 #FF0000 "
#> [124] "color /G:124 #FF0000 " "color /G:125 #FF0000 " "color /G:126 #FF0000 "
#> [127] "color /G:127 #FF0000 " "color /G:128 #FF0000 " "color /G:129 #FF0000 "
#> [130] "color /G:130 #FF0000 " "color /G:131 #FF0000 " "color /G:132 #FF0000 "
#> [133] "color /G:133 #000000 " "color /G:134 #FF0000 " "color /G:135 #FF0000 "
#> [136] "color /G:136 #684354 " "color /G:137 #714846 " "color /G:138 #714846 "
#> [139] "color /G:139 #714846 " "color /G:140 #714846 " "color /G:141 #714846 "
#> [142] "color /G:142 #714846 " "color /G:143 #5E257C " "color /G:144 #711B73 "
#> [145] "color /G:145 #711B73 " "color /G:146 #711B73 " "color /G:147 #000000 "
#> [148] "color /G:148 #711B73 " "color /G:149 #711B73 " "color /G:150 #711B73 "
#> [151] "color /G:151 #8D1B57 " "color /G:152 #8D1B57 " "color /G:153 #8D1B57 "
#> [154] "color /G:154 #433A82 " "color /G:155 #0D27CB " "color /G:156 #0D27CB "
#> [157] "color /G:157 #0D27CB " "color /G:158 #0D27CB " "color /G:159 #0D27CB "
#> [160] "color /G:160 #773355 " "color /G:161 #7F3051 " "color /G:162 #7F3051 "
#> [163] "color /G:163 #8E2749 " "color /G:164 #8E2749 " "color /G:165 #8E2749 "
#> [166] "color /G:166 #8D284A " "color /G:167 #92313B " "color /G:168 #92313B "
#> [169] "color /G:169 #92323B " "color /G:170 #874137 " "color /G:171 #874137 "
#> [172] "color /G:172 #874137 " "color /G:173 #874137 " "color /G:174 #874137 "
#> [175] "color /G:175 #864832 " "color /G:176 #864832 " "color /G:177 #724F3D "
#> [178] "color /G:178 #000000 " "color /G:179 #724F3D " "color /G:180 #724F3D "
#> [181] "color /G:181 #724F3D " "color /G:182 #695442 " "color /G:183 #625746 "
#> [184] "color /G:184 #625746 " "color /G:185 #375B6D " "color /G:186 #2F6170 "
#> [187] "color /G:187 #2F6170 " "color /G:188 #2F6170 " "color /G:189 #2F6170 "
#> [190] "color /G:190 #445E5D " "color /G:191 #746823 " "color /G:192 #746823 "
#> [193] "color /G:193 #746823 " "color /G:194 #7E6918 " "color /G:195 #8A6212 "
#> [196] "color /G:196 #8A6212 " "color /G:197 #92521B " "color /G:198 #92521B "
#> [199] "color /G:199 #92521B " "color /G:200 #92521B " "color /G:201 #92521B "
#> [202] "color /G:202 #556545 " "color /G:203 #0B8470 " "color /G:204 #03807C "
#> [205] "color /G:205 #028E6F " "color /G:206 #028E6F " "color /G:207 #028E6F "
#> [208] "color /G:208 #028C71 " "color /G:209 #0A866F " "color /G:210 #395D69 "
#> [211] "color /G:211 #4A3D79 " "color /G:212 #403F7F " "color /G:213 #403F7F "
#> [214] "color /G:214 #403F7F " "color /G:215 #403F7F " "color /G:216 #403F7F "
#> [217] "color /G:217 #555E4C " "color /G:218 #81532B " "color /G:219 #965C0E "
#> [220] "color /G:220 #965C0E " "color /G:221 #965C0E "