Create fits for all the peptides

This is a wrapper function for get_fit_results, but for multiple peptides at the same time. First, it creates a peptide list from the supplied deuterium uptake data, and calls the fitting function for each, to create full dataset.

Usage

create_fit_dataset(
  kin_dat,
  fit_k_params,
  control = list(maxiter = 1000, scale = "levenberg"),
  trace = F,
  fractional = F,
  workflow = 321
)

Arguments

kin_dat

...

fit_k_params

...

control

...

trace

...

fractional

...

workflow

...

Value

a data.frame object.

See also

Examples

kin_dat <- prepare_kin_dat(alpha_dat)
fit_k_params <- get_example_fit_k_params()
head(create_fit_dataset(kin_dat, fit_k_params))
#>   id   Protein       State    sequence start end max_uptake       n_1      k_1
#> 1  1 db_eEF1Ba ALPHA_Gamma GFGDLKSPAGL     1  11          9 2.5918362 5.870020
#> 2  2 db_eEF1Ba ALPHA_Gamma  FGDLKSPAGL     2  11          8 2.6495960 7.234888
#> 3  3 db_eEF1Ba ALPHA_Gamma   GDLKSPAGL     3  11          7 2.7013593 7.507887
#> 4  4 db_eEF1Ba ALPHA_Gamma      LKSPAG     5  10          4 1.3953257 6.717443
#> 5  5 db_eEF1Ba ALPHA_Gamma     LKSPAGL     5  11          5 1.5263186 6.866652
#> 6  6 db_eEF1Ba ALPHA_Gamma      AGLQVL     9  14          5 0.3919301 2.603586
#>        n_2       k_2      n_3          k_3          rss        bic class_name
#> 1 1.747477 0.1471729 1.093423 0.0178303882 1.972152e-31 -405.38705         NA
#> 2 1.671734 0.4732856 0.000000 0.0000000000 8.022972e-03  -13.71718         NA
#> 3 1.805400 0.4905359 0.000000 0.0000000000 2.168912e-03  -21.56567         NA
#> 4 1.356225 0.4680721 0.000000 0.0000000000 1.133739e-03  -25.45790         NA
#> 5 1.585653 0.4754323 0.000000 0.0000000000 1.075759e-03  -25.77287         NA
#> 6 2.271788 0.2740198 5.000000 0.0001269509 5.257664e-03  -12.66939         NA
#>       k_est fitted   color
#> 1 2.8513818      3 #7A5233
#> 2 4.6191186      2 #9C6300
#> 3 4.6967479      2 #996600
#> 4 3.6371609      2 #817E00
#> 5 3.6101125      2 #7D8200
#> 6 0.2144616      3 #0D4CA6