Creates uptake data from mass measurements

Function calculates deuterium uptake for each time point in the original data with respect to the parameters.

Usage

prepare_kin_dat(
  dat,
  state = dat[["State"]][1],
  time_0 = min(dat[["Exposure"]]),
  time_100 = max(dat[["Exposure"]]),
  replicate = F,
  FD = FALSE
)

Arguments

dat

experimental data read by HaDeX::read_hdx

state

biological state

time_0

minimal exchange control time point

time_100

maximal exchange control time point

replicate

idicator, if the replicate values should be provided, or aggregated

FD

idicator, if the time_100 value shoul be used as fully deuterated control, only for normalization purposes, or left as time point of measurement

Value

a data.frame object.

See also

Examples

kin_dat <- prepare_kin_dat(alpha_dat)
head(kin_dat)
#>   ID Exposure   Protein                     Sequence Start End       State
#> 1  1    0.167 db_eEF1Ba      GSSHHHHHHSSGLVPRGSHRSEF     2  24 ALPHA_Gamma
#> 2  2    0.167 db_eEF1Ba                       FMGFGD    24  29 ALPHA_Gamma
#> 3  3    0.167 db_eEF1Ba                FMGFGDLKSPAGL    24  36 ALPHA_Gamma
#> 4  4    0.167 db_eEF1Ba                        MGFGD    25  29 ALPHA_Gamma
#> 5  5    0.167 db_eEF1Ba MGFGDLKSPAGL+Oxidation M (1)    25  36 ALPHA_Gamma
#> 6  6    0.167 db_eEF1Ba              MGFGDLKSPAGLQVL    25  39 ALPHA_Gamma
#>   MaxUptake frac_deut_uptake err_frac_deut_uptake deut_uptake err_deut_uptake
#> 1        21        0.9525516          0.025707443    5.098016     0.012896475
#> 2         5        1.0154452          0.020267468    2.194507     0.023370098
#> 3        11        0.6114155          0.004053593    3.624016     0.016261556
#> 4         4        1.0095756          0.028675730    1.694090     0.009286482
#> 5        10        0.5926917          0.004760097    3.316276     0.005617954
#> 6        13        0.4183319          0.003182058    3.398111     0.017202403